# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/profiles/me": {
      "get": {
        "operationId": "Profiles_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/MyProfileDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ผู้เรียกในมุมของระบบแจ้งซ่อม (ชื่อจาก Core Hub · role ที่ใช้จริง · สิทธิ์ · รูป)",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles/me/avatar": {
      "post": {
        "operationId": "Profiles_uploadAvatar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "avatar"
                ],
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProfileDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT — ชนกฎธุรกิจ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "เปลี่ยนรูปโปรไฟล์ของตัวเอง (JPG · PNG · WebP ไม่เกิน 2 MB ในช่อง avatar)",
        "tags": [
          "profiles"
        ]
      },
      "delete": {
        "operationId": "Profiles_removeAvatar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeletedDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT — ชนกฎธุรกิจ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ลบรูปโปรไฟล์ของตัวเอง (กลับไปใช้อักษรย่อ)",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles": {
      "get": {
        "operationId": "Profiles_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "USER",
                "TECHNICIAN",
                "ADMIN"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProfileDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ผู้ที่เคยเข้าระบบนี้ตาม role (ใช้เลือกช่างตอนมอบหมายงาน)",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles/people": {
      "get": {
        "operationId": "Profiles_people",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "ค้นในรหัสและชื่อ (Core Hub)",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "personType",
            "required": false,
            "in": "query",
            "schema": {
              "default": "STUDENT",
              "type": "string",
              "enum": [
                "STUDENT",
                "STAFF"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PersonListItemDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "503": {
            "description": "SERVICE_UNAVAILABLE — ติดต่อ Core Hub ไม่ได้ชั่วคราว (มี header Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "รายชื่อนักศึกษา/บุคลากรจาก Core Hub + จำนวนใบแจ้งซ่อมในระบบนี้",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles/technicians/{personCode}": {
      "put": {
        "operationId": "Profiles_setTechnician",
        "parameters": [
          {
            "name": "personCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTechnicianDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT — ชนกฎธุรกิจ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "503": {
            "description": "SERVICE_UNAVAILABLE — ติดต่อ Core Hub ไม่ได้ชั่วคราว (มี header Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "แต่งตั้ง/ถอดถอนช่างซ่อมบำรุง (บุคลากรสายสนับสนุนจาก Core Hub)",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles/{id}/avatar": {
      "get": {
        "operationId": "Profiles_avatar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ไฟล์รูป",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "image/jpeg": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              },
              "image/png": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              },
              "image/webp": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "image/jpeg": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              },
              "image/png": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              },
              "image/webp": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "image/jpeg": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              },
              "image/png": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              },
              "image/webp": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "รูปโปรไฟล์ (image/jpeg · image/png · image/webp)",
        "tags": [
          "profiles"
        ]
      }
    },
...
  },
  "info": {
    "title": "csmju-maintenance-request API",
    "description": "ระบบแจ้งซ่อม — ระบบย่อยของ CSMJU2030 · ทุก endpoint ใต้ /api/v1 ต้องมี Core Hub access token (Authorization: Bearer หรือคุกกี้ csmju_maintenance_request_access_token) · response ห่อด้วย envelope มาตรฐาน",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT (RS256 จาก Core Hub)",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorCode": {
        "type": "string",
        "enum": [
          "BAD_REQUEST",
          "VALIDATION_ERROR",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "NOT_FOUND",
          "CONFLICT",
          "TOO_MANY_REQUESTS",
          "INTERNAL_ERROR",
          "SERVICE_UNAVAILABLE"
        ]
      },
      "ErrorBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              }
            ]
          },
          "message": {
            "type": "string",
            "description": "ข้อความภาษาไทยที่แสดงให้ผู้ใช้ได้ทันที"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
      "SessionDto": {
        "type": "object",
        "properties": {
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "token หมดอายุเมื่อไร — ต่ออายุล่วงหน้าผ่าน /auth/login"
          }
        },
        "required": [
          "expiresAt"
        ]
      },
      "MyProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "user-003",
            "description": "เท่ากับ claim `sub` ของ token"
          },
          "coreRole": {
            "type": "string",
            "enum": [
              "student",
              "alumni",
              "staff",
              "lecturer",
              "guest",
              "admin"
            ],
            "description": "เท่ากับ claim `role` ของ token"
          },
          "subsystemRole": {
            "type": "string",
            "enum": [
              "USER",
              "TECHNICIAN",
              "ADMIN"
            ],
            "description": "role ที่ใช้จริง (รวมการแต่งตั้งช่าง)"
          },
          "permissions": {
            "example": [
              "repair-request:create"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "personCode": {
            "type": "string",
            "nullable": true,
            "description": "รหัสนักศึกษา/บุคลากรจาก Core Hub (GET /people/me) · null = บัญชียังไม่ผูกกับบุคคลในทะเบียน"
          },
          "displayName": {
            "type": "string",
            "example": "นายสมชาย ใจดี",
            "description": "ชื่อจาก Core Hub (GET /people/me) · ดูไม่ได้ = รหัส"
          },
          "nameFromCoreHub": {
            "type": "boolean",
            "description": "false = ชื่อด้านบนไม่ได้มาจาก Core Hub"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "description": "รูปโปรไฟล์ (GET ได้เมื่อเข้าสู่ระบบแล้ว) · null = ยังไม่มีรูป ให้แสดงอักษรย่อแทน"
          },
          "session": {
            "$ref": "#/components/schemas/SessionDto"
          }
        },
        "required": [
          "id",
          "coreRole",
          "subsystemRole",
          "permissions",
          "personCode",
          "displayName",
          "nameFromCoreHub",
          "avatarUrl",
          "session"
        ]
      },
      "ProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "id ของโปรไฟล์ในระบบนี้"
          },
          "coreUserId": {
            "type": "string",
            "example": "user-003",
            "description": "claim `sub` จาก Core Hub"
          },
          "personCode": {
            "type": "string",
            "nullable": true,
            "example": "somsak.m",
            "description": "รหัสนักศึกษา/บุคลากรจาก Core Hub (GET /people/me) · null = บัญชียังไม่ผูกกับบุคคลในทะเบียน"
          },
          "coreRole": {
            "type": "string",
            "enum": [
              "student",
              "alumni",
              "staff",
              "lecturer",
              "guest",
              "admin"
            ]
          },
          "subsystemRole": {
            "type": "string",
            "enum": [
              "USER",
              "TECHNICIAN",
              "ADMIN"
            ],
            "nullable": true,
            "description": "null = core role นี้เข้าระบบไม่ได้แล้ว"
          },
          "isTechnician": {
            "type": "boolean"
          },
          "displayName": {
            "type": "string",
            "example": "somsak.m",
            "description": "รายการไม่หาชื่อจาก Core Hub ทีละแถว จึงเป็น personCode (reference-data.md ข้อ 7.2)"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "description": "รูปโปรไฟล์ (GET ได้เมื่อเข้าสู่ระบบแล้ว) · null = ยังไม่มีรูป ให้แสดงอักษรย่อแทน"
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "coreUserId",
          "personCode",
          "coreRole",
          "subsystemRole",
          "isTechnician",
          "displayName",
          "avatarUrl",
          "lastSeenAt"
        ]
      },
      "DeletedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "deleted": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "id",
          "deleted"
        ]
      },
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 42
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "PersonListItemDto": {
        "type": "object",
        "properties": {
          "personCode": {
            "type": "string",
            "example": "6504101234"
          },
          "fullName": {
            "type": "string",
            "example": "นางสาวกริญญา ทาเกร"
          },
          "personType": {
            "type": "string",
            "enum": [
              "STUDENT",
              "STAFF"
            ]
          },
          "staffType": {
            "type": "string",
            "nullable": true,
            "description": "ประเภทบุคลากร เช่น LECTURER"
          },
          "status": {
            "type": "string",
            "example": "ACTIVE"
          },
          "departmentName": {
            "type": "string",
            "nullable": true
          },
          "hasProfile": {
            "type": "boolean",
            "description": "เคยเข้าระบบแจ้งซ่อมแล้ว"
          },
          "isTechnician": {
            "type": "boolean"
          },
          "canBeTechnician": {
            "type": "boolean",
            "description": "แต่งตั้งเป็นช่างได้ (บุคลากรสายสนับสนุนที่มีบัญชีใน Core Hub)"
          },
          "requestCount": {
            "type": "number",
            "description": "ใบแจ้งซ่อมที่คนนี้แจ้งในระบบนี้"
          },
          "openRequestCount": {
            "type": "number",
            "description": "ที่ยังไม่ปิดงาน"
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "personCode",
          "fullName",
          "personType",
          "staffType",
          "status",
          "departmentName",
          "hasProfile",
          "isTechnician",
          "canBeTechnician",
          "requestCount",
          "openRequestCount",
          "lastSeenAt"
        ]
      },
      "SetTechnicianDto": {
        "type": "object",
        "properties": {
          "isTechnician": {
            "type": "boolean",
            "description": "แต่งตั้ง (true) / ถอดถอน (false) ช่างซ่อมบำรุง"
          }
        },
        "required": [
          "isTechnician"
        ]
      },
...
    }
  }
}